% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annots_primary_hg38}
\alias{annots_primary_hg38}
\alias{annotations}
\alias{annots_primary_mm10}
\alias{annots_full_hg38}
\alias{annots_full_mm10}
\title{Annotations}
\usage{
annots_primary_hg38(quiet = FALSE)

annots_full_hg38(quiet = FALSE)

annots_primary_mm10(quiet = FALSE)

annots_full_mm10(quiet = FALSE)
}
\arguments{
\item{quiet}{If TRUE, messages are suppressed. Default: FALSE.}
}
\value{
A list of \code{tbl} objects. See details.
}
\description{
Genomic features relevant to R-loops. Both mm10 and
hg38 annotations are available.
}
\details{
The \code{list} contains \code{tbl} objects (tidyverse-style data frames) containing
annotations as genomic ranges. The \strong{primary} annotations (e.g.,
\code{annots_primary_hg38()}) are an abbreviated
version of the \strong{full} annotations (e.g., \code{annots_full_hg38()}). See the
description below for further details:
\subsection{Databases available}{

This section details the annotation databases which are available in RLHub. See the succeeding
section ("Objects available based on accessor") for a list of which databases
are available within each function. All processing was performed using
\href{https://github.com/Bishop-Laboratory/RLBase-data/blob/e63707e9e4dece5b2c007bd4284975be4998dfd4/scripts/getGenomicFeatures.R}{this script}
as part of the \href{https://github.com/Bishop-Laboratory/RLBase-data}{RLBase-data} processing protocol.
\itemize{
\item \strong{Centromeres}
\itemize{
\item Description: Centromere locations within the genome.
\item Source: UCSC table \href{https://genome.ucsc.edu/cgi-bin/hgTables?db=hg38&hgta_group=map&hgta_track=centromeres&hgta_table=centromeres&hgta_doSchema=describe+table+schema}{centromeres}.
}
\item \strong{CNA}
\itemize{
\item Description: Copy-number alterations found in inherited disorder cell lines. See source for full description. CNV states (0-4)
are represented in the data as separate types. For example, Deep deletion (0) sites are accessed with \code{annots_full_hg38()$CNA__0}.
\item Source:  UCSC table \href{https://genome.ucsc.edu/cgi-bin/hgTables?db=hg38&hgta_group=phenDis&hgta_track=coriellDelDup&hgta_table=coriellDelDup&hgta_doSchema=describe+table+schema}{coriellDelDup}.
}
\item \strong{Cohesin}
\itemize{
\item Description: This database contains manually-curated STAG2 and STAG1 ChIP-Seq data reprocessed by the RLHub authors to find
consensus STAG1 and STAG2 sites between cell lines (using \href{https://github.com/rhysnewell/ChIP-R}{chip-r}.)
\item Source: processed data source \emph{\href{https://academic.oup.com/nar/article/48/10/5639/5827199}{Pan et al}} and processed
further as part of the RLBase-data pipeline \href{https://github.com/Bishop-Laboratory/RLBase-data/tree/e63707e9e4dece5b2c007bd4284975be4998dfd4}{here}.
}
\item \strong{CpG_Islands}:
\itemize{
\item Description: CpG island predicted locations throughout the genome.
\item Source: UCSC table \href{http://genome.ucsc.edu/cgi-bin/hgTables?hgta_doSchemaDb=hg38&hgta_doSchemaTable=cpgIslandExt}{cpgIslandExt}.
}
\item \strong{Encode_CREs}:
\itemize{
\item Description: The UCSC Encode_CREs table contains putative promoter-like ("prom"), promoter-enhancer-like ("enhP"),
distal-enhancer-like ("enhD"), H3K4me3 ("K4me3"), and CTCF ("CTCF") chromatin states across the genome.
\item Source: UCSC table \href{https://genome.ucsc.edu/cgi-bin/hgTables?db=hg38&hgta_group=regulation&hgta_track=encodeCcreCombined&hgta_table=encodeCcreCombined&hgta_doSchema=describe+table+schema}{encodeCcreCombined}.
}
\item \strong{Encode_Histone}:
\itemize{
\item Description: consensus peaks from histone ChIP experiments downloaded from Encode.
Biological replicates were summarized with \href{https://github.com/rhysnewell/ChIP-R}{chip-r}.
\item Source: Encode v121. Manifest of samples downloaded is \href{https://github.com/Bishop-Laboratory/RLBase-data/blob/e63707e9e4dece5b2c007bd4284975be4998dfd4/misc-data/histone_encode_manifest.csv}{here}.
}
\item \strong{encodeTFBS}:
\itemize{
\item Description: The collection of curated transcription-factor binding profiles from encode, made available by UCSC table browser.
\item Source: UCSC table \href{https://genome.ucsc.edu/cgi-bin/hgTrackUi?db=hg38&g=encRegTfbsClustered}{encRegTfbsClustered}.
}
\item \strong{G4Qexp}:
\itemize{
\item Description: G4-Quadruplex ChIP-Seq data
\item Source: GEO accession \href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE63874}{GSE63874}.
}
\item \strong{G4Qpred}:
\itemize{
\item Description: Re-processed and binned G4-Quadruplex Predictions. The type names for this database are the G4Q prediction classes and follow the pattern \code{tl:N_nl:N_gn:N}.
\code{tl}: the length of guanine tracts in region; \code{nl}: number of locations for G4 formation; \code{gn}: the number of possible simultaneous G4 structures. For more information, see the source publication \href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0165101}{here}.
Due to the large number of possible configurations of \code{tl}, \code{nl}, and \code{gn}, they were binned based on frequency.
\item Source: Figshare \href{https://figshare.com/collections/G4_Supporting_Material/3498270/1}{Rouchka et al.} and \href{https://figshare.com/ndownloader/files/6432597}{direct download link}
}
\item \strong{knownGene_RNAs}:
\itemize{
\item Description: RNA species provided by UCSC KnownGene, split up by the "transcriptType" column from the source table.
\item Source: UCSC table \href{http://genome.ucsc.edu/cgi-bin/hgTables?hgta_doSchemaDb=hg38&hgta_doSchemaTable=knownGene}{knownGene}.
}
\item \strong{Microsatellite}:
\itemize{
\item Description: Microsatellite DNA regions predicted based on motif.
\item Source: UCSC table \href{http://genome.ucsc.edu/cgi-bin/hgTables?hgta_doSchemaDb=hg38&hgta_doSchemaTable=microsat}{microsat}.
}
\item \strong{PolyA}:
\itemize{
\item Description: List of predicted poly-A sites, split by the "name2" column of the source table.
\item Source: UCSC table \href{http://genome.ucsc.edu/cgi-bin/hgTables?hgta_doSchemaDb=hg38&hgta_doSchemaTable=wgEncodeGencodePolyaV38}{wgEncodeGencodePolyaV38}.
}
\item \strong{RBP_ChIP}:
\itemize{
\item Description: ChIP-Seq data sets for RNA-binding proteins (RBPs) generated by \href{https://www.nature.com/articles/s41586-020-2077-3}{\emph{Nostrand et al}} for Encode. Data are split by ChIP target.
\item Source: Encode v121. Manifest of samples \href{https://static-content.springer.com/esm/art\%3A10.1038\%2Fs41586-020-2077-3/MediaObjects/41586_2020_2077_MOESM9_ESM.xlsx}{here} from source
study \href{https://www.nature.com/articles/s41586-020-2077-3}{\emph{Nostrand et al}}.
}
\item \strong{RBP_eCLiP}:
\itemize{
\item Description: eCLiP-Seq data sets for RNA-binding proteins (RBPs) generated by \href{https://www.nature.com/articles/s41586-020-2077-3}{\emph{Nostrand et al}} for Encode. Data are split by eCLiP target.
\item Source: Encode v121. Manifest of samples \href{https://static-content.springer.com/esm/art\%3A10.1038\%2Fs41586-020-2077-3/MediaObjects/41586_2020_2077_MOESM6_ESM.xlsx}{here} from source
study \href{https://www.nature.com/articles/s41586-020-2077-3}{\emph{Nostrand et al}}.
}
\item \strong{Repeat_Masker}:
\itemize{
\item Description: Repeat masker table from UCSC containing genomic annotations for predicted repetitive elements, split by class of repetitive element ("repClass").
\item Source: UCSC table \href{https://genome.ucsc.edu/cgi-bin/hgTables?db=hg38&hgta_table=rmsk&hgta_doSchema=describe+table+schema}{rmsk}.
}
\item \strong{skewr}:
\itemize{
\item Description: Regions of G or C-skew profiled using the \code{skewr} program. See the RLBase-data \href{https://github.com/Bishop-Laboratory/RLBase-data}{\code{README.md}} for steps.
\item Source: From UCSC goldenPath, \href{ftp://hgdownload.soe.ucsc.edu/goldenPath/hg38/bigZips/hg38.fa.gz}{hg38} and
\href{ftp://hgdownload.soe.ucsc.edu/goldenPath/mm10/bigZips/mm10.fa.gz}{mm10} genomes.
\href{ftp://hgdownload.soe.ucsc.edu/goldenPath/hg38/bigZips/genes/hg38.ensGene.gtf.gz}{hg38}
and \href{ftp://hgdownload.soe.ucsc.edu/goldenPath/mm10/bigZips/genes/mm10.ensGene.gtf.gz}{mm10} gene GTF.
CpG islands for mm10 and hg38 provided as described in the \strong{CpG_Islands} entry above. Processing
proceeded using \href{https://github.com/srhartono/SkewR/tree/e351e20c9d5a1704e7b5d79ad7dbdc895d04919d}{\code{skewr}} with
\href{https://github.com/KorfLab/StochHMM/releases/tag/v0.38}{\verb{stochHMM v0.38}}.
}
\item \strong{snoRNA_miRNA_scaRNA}:
\itemize{
\item Description: snoRNA, miRNA, and scaRNA species provided by UCSC table browser and split by the "type" column.
\item Source: UCSC table \href{http://genome.ucsc.edu/cgi-bin/hgTables?hgta_doSchemaDb=hg38&hgta_doSchemaTable=wgRna}{wgRna}.
}
\item \strong{Splice_Events}:
\itemize{
\item Description: UCSC table of alternative splice events predicted from transcriptome data sets. Split by "name" column.
\item Source: UCSC table \href{http://genome.ucsc.edu/cgi-bin/hgTables?hgta_doSchemaDb=hg38&hgta_doSchemaTable=knownAlt}{knownAlt}.
}
\item \strong{Transcript_Features}:
\itemize{
\item Description: Transcript features (e.g., "exon", "intron", etc) provided by Bioconductor TxDb packages. Split
based on the following features: "Exon", "Intron", "fiveUTR", "threeUTR", "TSS", "TTS", "Intergenic".
\item Source: TxDb for \href{https://bioconductor.org/packages/release/data/annotation/html/TxDb.Hsapiens.UCSC.hg38.knownGene.html}{hg38} and
\href{https://bioconductor.org/packages/release/data/annotation/html/TxDb.Mmusculus.UCSC.mm10.knownGene.html}{mm10}.
}
\item \strong{tRNAs}:
\itemize{
\item Description: UCSC table containing predicted tRNA genes.
\item Source: UCSC table \href{http://genome.ucsc.edu/cgi-bin/hgTables?hgta_doSchemaDb=hg38&hgta_doSchemaTable=tRNAs}{tRNAs}.
}
}
}

\subsection{Objects available based on accessor function}{

Here, we show which objects are available with each accessor function:\tabular{lllll}{
   DataBase name \tab \code{annots_primary_hg38()} \tab \code{annots_primary_mm10()} \tab \code{annots_full_hg38()} \tab \code{annots_full_mm10()} \cr
   Centromeres \tab FALSE \tab FALSE \tab TRUE \tab FALSE \cr
   CNA \tab FALSE \tab FALSE \tab TRUE \tab FALSE \cr
   Cohesin \tab FALSE \tab FALSE \tab TRUE \tab FALSE \cr
   CpG_Islands \tab TRUE \tab TRUE \tab TRUE \tab TRUE \cr
   Encode_CREs \tab TRUE \tab TRUE \tab TRUE \tab TRUE \cr
   Encode_Histone \tab FALSE \tab FALSE \tab TRUE \tab FALSE \cr
   encodeTFBS \tab FALSE \tab FALSE \tab TRUE \tab FALSE \cr
   G4Qexp \tab FALSE \tab FALSE \tab TRUE \tab FALSE \cr
   G4Qpred \tab TRUE \tab FALSE \tab TRUE \tab FALSE \cr
   knownGene_RNAs \tab TRUE \tab FALSE \tab TRUE \tab FALSE \cr
   Microsatellite \tab FALSE \tab FALSE \tab TRUE \tab TRUE \cr
   PolyA \tab TRUE \tab FALSE \tab TRUE \tab FALSE \cr
   RBP_ChIP \tab FALSE \tab FALSE \tab TRUE \tab FALSE \cr
   RBP_eCLiP \tab FALSE \tab FALSE \tab TRUE \tab FALSE \cr
   Repeat_Masker \tab TRUE \tab TRUE \tab TRUE \tab TRUE \cr
   skewr \tab TRUE \tab TRUE \tab TRUE \tab TRUE \cr
   snoRNA_miRNA_scaRNA \tab TRUE \tab FALSE \tab TRUE \tab FALSE \cr
   Splice_Events \tab FALSE \tab FALSE \tab TRUE \tab TRUE \cr
   Transcript_Features \tab TRUE \tab TRUE \tab TRUE \tab TRUE \cr
   tRNAs \tab TRUE \tab TRUE \tab TRUE \tab TRUE \cr
}

}

\subsection{Object structure}{

Accessor functions (e.g., \code{annots_primary_hg38()}) return a named \code{list} of \code{tbl}
objects that specify feature ranges. Below, we detail the naming and structure of each.
\subsection{List names}{

The names in the \code{list} objects provided by each accessor function (e.g., \code{annots_primary_hg38()})
follow this structure: \code{DataBase__Type}. \code{DataBase} is the database from which
annotations were derived and \code{Type} indicates the specific annotations from the database
which are included in the \code{tbl}. This is required as some databases produce > 1
type of annotation (e.g., \strong{Transcript_Features} contains "Exon"
(\code{Transcript_Features__Exon}) and "Intron" (\code{Transcript_Features__Intron})).
}

\subsection{\code{tbl} structure}{

Each \code{tbl} returned has the following structure:\tabular{lllll}{
   chrom \tab start \tab end \tab strand \tab id \cr
   chr1 \tab 10015 \tab 10498 \tab + \tab 1 \cr
   chr1 \tab 10614 \tab 11380 \tab + \tab 2 \cr
   ... \tab  \tab  \tab  \tab  \cr
}


Columns:
\itemize{
\item "chom" - the Chromosome of the feature range (UCSC style)
\item "start" - the starting position of the feature range.
\item "end" - the end position of the feature range.
\item "strand" - the strand of the feature range.
\item "id" - A unique ID for the feature range.
}
}

}
}
\examples{

annos <- annots_primary_hg38()

annos <- annots_full_hg38()

annos <- annots_primary_mm10()

annos <- annots_full_mm10()

}
