% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureEnrichment.R
\name{feat_enrich_samples}
\alias{feat_enrich_samples}
\alias{feat_enrich_rlregions}
\title{Feature Enrichment}
\usage{
feat_enrich_samples(quiet = FALSE)

feat_enrich_rlregions(quiet = FALSE)
}
\arguments{
\item{quiet}{If TRUE, messages are suppressed. Default: FALSE.}
}
\value{
A \code{tbl} object. See details.
}
\description{
Feature Enrichment per RL-Region (\link{feat_enrich_rlregions}) and per Sample (\link{feat_enrich_samples}).
}
\details{
\subsection{Source}{

For every R-loop-mapping sample in RLBase (\link{feat_enrich_samples}) or for
each of the three RL-Regions databases (\link{feat_enrich_rlregions}),
peaks were calculated and then
tested via \code{\link[RLSeq:featureEnrich]{RLSeq::featureEnrich()}} to determine enrichment within all
available genomic features.

The processing steps are described in the \href{https://github.com/Bishop-Laboratory/RLBase-data}{RLBase-data} README.md.
}

\subsection{Structure}{

Example data:\tabular{llrrrrrrrrrrrl}{
   db \tab type \tab num_total_peaks \tab num_tested_peaks \tab num_tested_anno_ranges \tab num_total_anno_ranges \tab avg_reldist_rl \tab avg_reldist_shuf \tab pval_reldist \tab stat_fisher_rl \tab pval_fisher_rl \tab stat_fisher_shuf \tab pval_fisher_shuf \tab experiment \cr
   Cohesin \tab STAG1 \tab 9892 \tab 10000 \tab 9935 \tab 9939 \tab 0.2514476 \tab 0.2480947 \tab 0.0816731 \tab 1.400271 \tab 0.0477869 \tab 0.6820876 \tab 0.1303705 \tab ERX2277510 \cr
   Cohesin \tab STAG2 \tab 10613 \tab 10000 \tab 9935 \tab 10696 \tab 0.2432965 \tab 0.2525422 \tab 0.0002326 \tab 1.945212 \tab 0.0000159 \tab 0.5382866 \tab 0.0174814 \tab ERX2277510 \cr
   CpG_Islands \tab CpG_Islands \tab 28189 \tab 10000 \tab 9989 \tab 31144 \tab 0.2190449 \tab 0.2501005 \tab 0.0000000 \tab 5.332338 \tab 0.0000000 \tab 0.9855223 \tab 0.9236551 \tab ERX2277510 \cr
   ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \cr
}


Column description:
\itemize{
\item \code{db} - the database from which annotations were derived. See \link{annotations}.
\item \code{type} - the type from which annotations were derived. See \link{annotations}.
\item \code{num_tested_peaks}: The number of peak ranges tested for enrichment.
\item \code{num_total_peaks}: The total number of peaks in the sample.
\item \code{num_tested_anno_ranges}: The number of annotation ranges which the peaks were tested against.
\item \code{num_total_anno_ranges}: The total number of ranges in the annotation.
\item \code{avg_reldist_rl}: The mean relative distances between peaks and annotation ranges as described \href{https://rnabioco.github.io/valr/reference/bed_reldist.html}{here}.
\item \code{avg_reldist_shuf}: Same as \code{avg_reldist_rl} except with peaks shuffled randomly to provide a negative control.
\item \code{pval_reldist}: The \code{ks.test} p value from comparing the distribution of the peak and shuffle relative distances as described \href{https://rnabioco.github.io/valr/articles/interval-stats.html}{here}.
\item \code{stat_fisher_rl}: The the odds ratio from the fisher's exact test as described \href{https://rnabioco.github.io/valr/reference/bed_fisher.html}{here}.
\item \code{stat_fisher_shuf}: Same as above, but with the shuffled peaks to provide a negative control.
\item \code{pval_fisher_rl}: The p value from Fisher's exact test as described \href{https://rnabioco.github.io/valr/reference/bed_fisher.html}{here}.
\item \code{pval_fisher_shuf}: Same as above, but with the shuffled peaks to provide a negative control.
\item \code{experiment} (only \code{feat_enrich_samples()}): The R-loop mapping sample for which the calculation was performed. See \link{rlbase_samples}.
\item \code{opt} (only \code{feat_enrich_rlregions()}): The type of RL-Region (see \link{rlregions}) which was tested.
}
}
}
\examples{

fes <- feat_enrich_samples()

fes <- feat_enrich_rlregions()
}
