% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_exp.R
\name{gene_exp}
\alias{gene_exp}
\title{Gene Expression}
\usage{
gene_exp(quiet = FALSE)
}
\arguments{
\item{quiet}{If TRUE, messages are suppressed. Default: FALSE.}
}
\value{
A \code{SummarizedExperiment} object. See details.
}
\description{
A \code{SummarizedExperiment} containing the gene expression results from RNA-Seq
data sets published in tandem with R-loop-mapping data sets.
}
\details{
\subsection{Source}{

For each R-loop-mapping data set in RLBase, any associated RNA-Sequencing data
was also downloaded and quantified using \verb{salmon v1.4.0}. This enables comparison
of R-loop abundance and gene expression. The list of SRA accessions from which
data was obtained are found here: \code{colData(gene_exp())$sample}.

The processing steps are described in the \href{https://github.com/Bishop-Laboratory/RLBase-data}{RLBase-data} README.md.
}

\subsection{Structure}{
\itemize{
\item colData
\itemize{
\item Contains metadata about every sample matching the metadata. See \link{rlbase_samples} for a full description.
}
\item assays
\itemize{
\item A \code{SimpleAssays} object containing three matrices (genes X samples):
\itemize{
\item \code{cts} - Raw read counts
\item \code{tpm} - TPM-normalized read counts
\item \code{vst} - VST-transformed read counts. (See \code{\link[DESeq2:vst]{DESeq2::vst()}})
}
}
}
}
}
\examples{

geneExp <- gene_exp()
}
