% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlbase_samples.R
\name{rlbase_samples}
\alias{rlbase_samples}
\title{RLBase Sample Manifest}
\usage{
rlbase_samples(quiet = FALSE)
}
\arguments{
\item{quiet}{If TRUE, messages are suppressed. Default: FALSE.}
}
\value{
A \code{tbl}.
}
\description{
A \code{tbl} containing metadata about each sample in RLBase.
}
\details{
\subsection{Source}{

RLBase samples were curated by hand in Excel from searching keywords such as "R-loops"
and "RNA:DNA hybrids" in GEO, SRA, and PubMed. Where R-loop mapping data
was publically available, entries were added to the Excel spreadsheet such that
every sample (SRX.../ERX.../GSM...) had it's own line. Information was noted for each
sample, such as the "mode" (the type of R-loop mapping it was) and the "Condition"
(e.g., "RNaseH1", "WKKD", etc). When genomic input controls were available, they
were manually matched to the experimental samples for which they could serve as a
background control during peak calling.

The up-to-date excel sheet is found \href{https://github.com/Bishop-Laboratory/RLBase-data/blob/main/rlbase-data/rlbase_catalog.xlsx}{here}.

Throughout the process of analyzing the data (see \href{https://github.com/Bishop-Laboratory/RLBase-data}{RLBase-data}), additional
metadata was added to the sample sheet (see structure for full account).
}

\subsection{Structure}{

\code{rlbase_samples} is a \code{tbl} with the structure:\tabular{llllllllrllllllllllllrlrlllrlllllllll}{
   rlsample \tab label \tab condition \tab mode \tab lab \tab tissue \tab genotype \tab other \tab PMID \tab group \tab family \tab ip_type \tab strand_specific \tab moeity \tab bisulfite_seq \tab file_type \tab experiment_original \tab control_original \tab study \tab name \tab paired_end \tab read_length \tab control \tab eff_genome_size \tab genome \tab prediction \tab discarded \tab numPeaks \tab expsamples \tab exp_matchCond \tab coverage_s3 \tab peaks_s3 \tab fastq_stats_s3 \tab bam_stats_s3 \tab report_html_s3 \tab rlranges_rds_s3 \tab rlfs_rda_s3 \cr
   SRX1070676 \tab POS \tab S96 \tab DRIPc \tab Fred Chedin \tab NT2 \tab WT \tab NT \tab 27373332 \tab rl \tab DRIP \tab S9.6 \tab TRUE \tab RNA \tab FALSE \tab public \tab GSM1720613 \tab NA \tab SRP059800 \tab GSM1720613: NT2 DRIPc-seq, rep 1; Homo sapiens; OTHER \tab FALSE \tab 50 \tab NA \tab 2706186140 \tab hg38 \tab POS \tab FALSE \tab 34092 \tab SRX1070685,SRX1070686 \tab WT_NT_SRP059800_NT2 \tab coverage/SRX1070676_hg38.bw \tab peaks/SRX1070676_hg38.broadPeak \tab fastq_stats/SRX1070676_hg38__fastq_stats.json \tab bam_stats/SRX1070676_hg38__bam_stats.txt \tab reports/SRX1070676_hg38.html \tab rlranges/SRX1070676_hg38.rds \tab rlfs_rda/SRX1070676_hg38.rlfs.rda \cr
   SRX1070677 \tab POS \tab S96 \tab DRIPc \tab Fred Chedin \tab NT2 \tab WT \tab NT \tab 27373332 \tab rl \tab DRIP \tab S9.6 \tab TRUE \tab RNA \tab FALSE \tab public \tab GSM1720614 \tab NA \tab SRP059800 \tab GSM1720614: NT2 DRIPc-seq, rep 2; Homo sapiens; OTHER \tab FALSE \tab 50 \tab NA \tab 2706186140 \tab hg38 \tab POS \tab FALSE \tab 22117 \tab SRX1070685,SRX1070686 \tab WT_NT_SRP059800_NT2 \tab coverage/SRX1070677_hg38.bw \tab peaks/SRX1070677_hg38.broadPeak \tab fastq_stats/SRX1070677_hg38__fastq_stats.json \tab bam_stats/SRX1070677_hg38__bam_stats.txt \tab reports/SRX1070677_hg38.html \tab rlranges/SRX1070677_hg38.rds \tab rlfs_rda/SRX1070677_hg38.rlfs.rda \cr
   SRX1070678 \tab POS \tab S96 \tab DRIP \tab Fred Chedin \tab NT2 \tab WT \tab NT \tab 27373332 \tab rl \tab DRIP \tab S9.6 \tab FALSE \tab DNA \tab FALSE \tab public \tab GSM1720615 \tab NA \tab SRP059800 \tab GSM1720615: NT2 DRIP-seq, 1; Homo sapiens; OTHER \tab FALSE \tab 50 \tab NA \tab 2706186140 \tab hg38 \tab POS \tab FALSE \tab 73924 \tab SRX1070685,SRX1070686 \tab WT_NT_SRP059800_NT2 \tab coverage/SRX1070678_hg38.bw \tab peaks/SRX1070678_hg38.broadPeak \tab fastq_stats/SRX1070678_hg38__fastq_stats.json \tab bam_stats/SRX1070678_hg38__bam_stats.txt \tab reports/SRX1070678_hg38.html \tab rlranges/SRX1070678_hg38.rds \tab rlfs_rda/SRX1070678_hg38.rlfs.rda \cr
   ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \cr
}


Column description:
\itemize{
\item \code{rlsample} - The unique ID of the sample, same as in the \href{https://www.ncbi.nlm.nih.gov/sra}{SRA}.
\item \code{label} - Label corresponding to the author-supplied condition of the sample. "POS" indicates the sample should robustly map R-loops, "NEG" indicates the opposite.
\item \code{condition} - The specific condition for each sample.
\item \code{mode} - The type of R-loop mapping for each sample.
\item \code{lab} - The senior author on the publication from which the data was provided.
\item \code{tissue} - The tissue condition for the sample.
\item \code{genotype} - The sample's genotype.
\item \code{other} - A column for other pertinent metadata provided by the authors.
\item \code{PMID} - The PMID associated with the sample
\item \code{group} - One of "rl" (R-loop mapping) or "exp" (Expression data/RNA-Seq).
\item \code{family} - The family of the "mode" (e.g., "DRIP" includes "sDRIP", "DRIPc", "qDRIP", etc)
\item \code{ip_type} - The IP type of the "mode" for each sample. One of "S9.6", "dRNH" (dead RNaseH1), or "None".
\item \code{strand_specific} - Whether the sample is stranded.
\item \code{moeity} - The moeity which was IP'd (if applicable)
\item \code{bisulfite_seq} - Whether the data uses bisulfite conversion sequencing (e.g., "BisDRIP-Seq" samples)
\item \code{file_type} - The type of data (always "public" for RLBase samples).
\item \code{experiment_original} - The original name of this sample as entered by hand in the curated Excel spreadsheet (usually converted from GSM to SRX).
\item \code{control_original} - Same as above for the accompanying control sample (if applicable.)
\item \code{study} - The SRA study accession for this sample.
\item \code{name} - The sample's name as entered in SRA.
\item \code{paired_end} - A logical indicating whether the data is paired end.
\item \code{read_length} - The read length for the sample.
\item \code{control} - The RLBase ID of the genomic input control sample corresponding to this sample (if applicable)
\item \code{eff_genome_size} - The effective genome size based on read length and genome (calculated with the \code{khmer} package) see relevant portion of RLBase-data protocol \href{https://github.com/Bishop-Laboratory/RLBase-data#preliminary}{here}.
\item \code{genome} - The UCSC genome ID for this sample.
\item \code{prediction} - The prediction from running \code{\link[RLSeq:predictCondition]{RLSeq::predictCondition()}}.
\item \code{discarded} - A logical indicating whether this sample was discarded during model building for a mismatch with its "label" (see \link{models}).
\item \code{numPeaks} - The number of peaks called for this sample.
\item \code{expsamples} - The IDs of any corresponding expression samples.
\item \code{exp_matchCond} - The meta data used to match this sample to any corresponding expression samples (if applicable)
\item \code{coverage_s3} - The location of the coverage tracks (\code{.bw}) in the AWS S3 bucket for RLBase data ('s3://rlbase_data/').
\item \code{peaks_s3} - Same as above for peak files (\code{.broadPeak})
\item \code{fastq_stats_s3} - Same as above for fastq QC statistics data (\code{.json}).
\item \code{bam_stats_s3} - Same as above for BAM QC statistics data (\code{.txt}).
\item \code{report_html_s3} - Same as above for reports from \code{\link[RLSeq:report]{RLSeq::report()}} (\code{.html}).
\item \code{rlranges_rds_s3} - Same as above for RLRanges R objects, as in \code{\link[RLSeq:RLRanges]{RLSeq::RLRanges()}} (\code{.rds})
\item \code{rlfs_rda_s3} - Same as above for \link{rlfs_res} objects generated by \code{\link[RLSeq:analyzeRLFS]{RLSeq::analyzeRLFS()}} (\code{.rda}).
}
}
}
\examples{

rlsamples <- rlbase_samples()

}
