% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlbps.R
\name{rlbps}
\alias{rlbps}
\title{R-loop Binding Proteins}
\usage{
rlbps(quiet = FALSE)
}
\arguments{
\item{quiet}{If TRUE, messages are suppressed. Default: FALSE.}
}
\value{
A \code{tbl} object. See details.
}
\description{
A list of annotations for genomic features relevant to R-loops. Both mm10 and
hg38 annotations are available.
}
\details{
\subsection{Source}{

R-Loop binding proteins (RLBPs) are proteins which were discovered by mass-spectrometry
after S9.6 immuno-precipitation (S9.6 is an antibody that binds R-loops). The data
were aggregated from three studies which analyzed RLBPs in this manner:
\enumerate{
\item \emph{Nadel et al., 2015} - \href{https://epigeneticsandchromatin.biomedcentral.com/articles/10.1186/s13072-015-0040-6}{article}
\item \emph{Wang et al., 2018} - \href{https://genome.cshlp.org/content/early/2018/08/14/gr.237362.118}{article}
\item \emph{Cristini et al., 2018} - \href{https://pubmed.ncbi.nlm.nih.gov/29742442/}{article}
}

The data were downloaded from the supplementary materials of each study
and processed using \href{https://github.com/Bishop-Laboratory/RLBase-data/tree/main/scripts/rlbps/}{custom R scripts}
as part of the \href{https://github.com/Bishop-Laboratory/RLBase-data}{RLBase-data} pipeline. This processing
involved normalization of mass-spec abundance calculations and the derivation of a "combined score"
for ranking all the proteins.
}

\subsection{Structure}{\tabular{lllll}{
   geneName \tab NADEL.2015 \tab WANG.2018 \tab CRISTINI.2018 \tab combinedScore \cr
   DDX21 \tab 0.9 \tab 0 \tab 0.968 \tab 1.87 \cr
   DHX9 \tab 0.85 \tab 0.0635 \tab 0.831 \tab 1.74 \cr
   ... \tab ... \tab ... \tab ... \tab ... \cr
}


Columns:
\itemize{
\item geneName
\itemize{
\item The official HGNC symbol of the RLBP
}
\item NADEL.2015
\itemize{
\item Normalized abundance of protein from \emph{Nadel et al., 2015}
}
\item WANG.2018
\itemize{
\item Normalized abundance of protein from \emph{Wang et al., 2018}
}
\item CRISTINI.2018
\itemize{
\item Normalized abundance of protein from \emph{Cristini et al., 2018}
}
\item combinedScore
\itemize{
\item Sum of normalized abundance across studies.
}
}
}
}
\examples{

rlbpsData <- rlbps()

}
