% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlfs_res.R
\name{rlfs_res}
\alias{rlfs_res}
\title{RLFS-Test Results}
\usage{
rlfs_res(quiet = FALSE)
}
\arguments{
\item{quiet}{If TRUE, messages are suppressed. Default: FALSE.}
}
\value{
A named \code{list}.
}
\description{
A list containing the results of R-loop-forming sequences analysis (\code{\link[RLSeq:analyzeRLFS]{RLSeq::analyzeRLFS()}}) and
subsequent sample classification \code{\link[RLSeq:predictCondition]{RLSeq::predictCondition()}} as "POS"
(robustly maps R-loops) or "NEG" (poorly maps R-loops) with all samples
in RLBase.
}
\details{
\subsection{Source}{

R-loop-forming sequences were computationally predicted using \href{https://github.com/piroonj/QmRLFS-finder}{\code{QmRLFS-finder.py}}. The relevant section
of the RLBase data prep protocol can be found \href{https://github.com/Bishop-Laboratory/RLBase-data#generate-datasets}{here}.
Briefly, genomes available from UCSC were downloaded and analyzed with \code{QmRLFS-finder.py}
on default settings. The resulting RLFS ranges were converted to \code{.bed} format.

RLFS analysis (via \code{\link[RLSeq:analyzeRLFS]{RLSeq::analyzeRLFS()}}) implements permutation testing to calculate
the enrichment of a query R-loop mapping peakset within the relevant RLFS ranges. This analysis produces
a p value and a Z-score distribution which are subsequently analyzed with
\code{\link[RLSeq:predictCondition]{RLSeq::predictCondition()}} (see \link{models}), yielding a quality prediction of "POS"
(robustly maps R-loops) or "NEG" (poorly maps R-loops).

This approach was applied to every sample in RLBase (see \link{rlbase_samples}) to yield \link{rlfs_res}.
}

\subsection{Structure}{

\code{rlfs_res} is a named list where each name is the ID of a sample in RLBase (see \link{rlbase_samples}).

Each \code{list} item is itself a \code{list} comprising the following:
\itemize{
\item \code{rlfsData} - A \code{list} with this structure:
\itemize{
\item \code{perTestResults}
\itemize{
\item An object of the class \code{permTestResultsList} from \code{regioneR}.
\item Contains the results of permutation testing.
}
\item \code{Z-scores}
\itemize{
\item An object of the class \code{localZScoreResultsList} from \code{regioneR}.
\item Contains the results of local Z-score analysis +/-5kb around each RLFS.
}
}
\item \code{rlfsPred} - A \code{list} with this structure:
\itemize{
\item \code{Features}
\itemize{
\item A \code{tbl} containing the raw and transformed values of features used in \code{\link[RLSeq:predictCondition]{RLSeq::predictCondition()}}.
}
\item \code{Criteria}
\itemize{
\item A \code{list} containing the results of analyzing the four criteria which must all be \code{TRUE} for a prediction of "POS":
\itemize{
\item \verb{PVal Significant} - The permutation test p value is < .05
\item \code{ZApex > 0} - The center of the Z-score distribution is > 0.
\item \code{ZApex > ZEdge} - The center of the Z-score distribution is > left edge and right edge.
\item \verb{Predicted 'Case'} - The classifier predicted the "Case" label ("POS").
}
}
\item \code{prediction}
\itemize{
\item The final prediction: "POS" or "NEG"
}
}
}
}
}
\examples{

rlfsRes <- rlfs_res()

}
