% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlregions.R
\name{rlregions_annot}
\alias{rlregions_annot}
\alias{rlregions}
\alias{rlregions_meta}
\alias{rlregions_counts}
\title{R-loop Regions}
\usage{
rlregions_annot(quiet = FALSE)

rlregions_meta(quiet = FALSE)

rlregions_counts(quiet = FALSE)
}
\arguments{
\item{quiet}{If TRUE, messages are suppressed. Default: FALSE.}
}
\value{
A list of \code{tbl} objects. See details.
}
\description{
R-loop regions (RL regions) are consensus sites of R-loop formation derived
from a meta-analysis of the R-loop mapping experiments in RLBase. RLHub includes
information about these regions (\link{rlregions_meta}), the overlap of genomic
features with RL regions (\link{rlregions_annot}), and the read count matrices
for each sample in RLBase within each RL region (\link{rlregions_counts}).
}
\details{
R-loop regions (RL regions) are consensus sites of R-loop formation derived
from a meta-analysis of the R-loop mapping experiments in RLBase. In \code{RLHub},
we provide access to:
\itemize{
\item \link{rlregions_meta} - The metadata describing these RL regions.
\item \link{rlregions_annot} - RL regions overlapped with genome features (see \link{annotations}).
\item \link{rlregions_counts} - RL region read counts across RLBase samples.
}
}
\section{Source}{
\subsection{RL Regions - \link{rlregions_meta}}{

RL Regions were derived during \href{https://github.com/Bishop-Laboratory/RLBase-data#get-r-loop-consensus}{this}
step of the RLBase-data \href{https://github.com/Bishop-Laboratory/RLBase-data}{protocol}.

Here is a brief summary of the key processing steps.
\enumerate{
\item Human RLBase samples which were labeled as "POS", classified as "POS" (see \link{rlfs_res}), and which
have at least 5000 peaks were selected.
\item Then, these peaks were then randomly downsampled to 5000 ranges each and aggregated.
\item Then, the hg38 genome was binned into 10bp bins and the number of aggregated peaks was
overlapped with each bin and counted, producing a bedGraph.
\item Then, macs3 was implemented to call peaks on the bedGraph file.
\item Finally, the resulting ranges were combined to create summary ranges.
}

\strong{Note}: steps 2-4 were performed separately on datasets from dRNH (catalytically-dead RNaseH1) R-loop mapping experiments
and S9.6-based experiments due to the noted differences in R-loop mapping between these modalities. In step 5, they were combined to
create summary ranges, and the original source (dRNH or S9.6) was noted in the resulting metadata.
}

\subsection{RL Regions anno - \link{rlregions_annot}}{

Once RL regions were generated, they were overlapped with the annotations found in \code{\link[=annots_full_hg38]{annots_full_hg38()}}.
This was done separately for dRNH-derived, S9.6-derived, and combined RL regions.
}

\subsection{RL Regions Counts - \link{rlregions_counts}}{

Once RL regions were generated, \code{\link[Rsubread:featureCounts]{Rsubread::featureCounts()}} was used to count
the number of overlapping reads from the \code{.bam} alignment files for each
sample in RLBase (\link{rlbase_samples}) within each RL region (only "combined" regions used).

These data were then VST-transformed and TPM was also calculated.
}
}

\section{Structure}{
\subsection{RL Regions - \link{rlregions_meta}}{

\code{rlregions_meta()} is a \code{tbl} with the following structure:\tabular{llllrrrrrrrrrrrrlrrrrllll}{
   rlregion \tab location \tab is_rlfs \tab source \tab confidence_level \tab medSignalVal \tab medPVal \tab medQVal \tab avgSignalVal \tab avgPVal \tab avgQVal \tab numPeaks \tab nSamples \tab nStudies \tab nModes \tab nTissues \tab ip_types \tab nIPTypes \tab pct_case \tab avgNumPeaks \tab medNumPeaks \tab geneIDs \tab allGenes \tab mainGenes \tab is_repeat \cr
   All_RL64926 \tab chrM:0-16569:. \tab TRUE \tab dRNH S96 \tab 53.21 \tab 1.962881 \tab 20.92420 \tab 17.01782 \tab 2.563848 \tab 653.89461 \tab 649.8001 \tab 52373 \tab 298 \tab 44 \tab 16 \tab 36 \tab S9.6,dRNH,RNA-m6A,None \tab 4 \tab 75.16779 \tab 80117.79 \tab 45063 \tab NA \tab NA \tab NA \tab TRUE \cr
   All_RL19159 \tab chr5:49656610-49661950:. \tab FALSE \tab dRNH S96 \tab 36.55 \tab 2.826610 \tab 19.16160 \tab 15.24025 \tab 3.282115 \tab 108.41481 \tab 104.6716 \tab 107393 \tab 261 \tab 43 \tab 15 \tab 34 \tab S9.6,RNA-m6A,None,dRNH \tab 4 \tab 77.01149 \tab 73058.96 \tab 43815 \tab NA \tab NA \tab NA \tab TRUE \cr
   All_RL50251 \tab chr16:46386300-46391280:. \tab FALSE \tab dRNH S96 \tab 31.68 \tab 3.258560 \tab 25.21965 \tab 21.00418 \tab 3.428755 \tab 87.33233 \tab 83.3639 \tab 133750 \tab 243 \tab 43 \tab 16 \tab 33 \tab RNA-m6A,S9.6,dRNH,None \tab 4 \tab 76.13169 \tab 65721.60 \tab 38878 \tab NA \tab NA \tab NA \tab TRUE \cr
   ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \cr
}


Column description:
\itemize{
\item \code{rlregion} - The ID of the RL Region.
\item \code{location} - The location of the RL Region. This column follows the pattern "chrom:start-stop:strand"
\item \code{is_rlfs} - Logical indicating whether the RL region overlaps with an R-loop forming sequence.
\item \code{source} - Indicates whether the RL region was discovered from catalytically-dead RNaseH1-based mapping (\code{dRNH}), S9.6-based mapping (\code{S96}), or both (\verb{dRNH S96}).
\item \code{confidence_level} - The percent of samples in which this range was found.
\item \code{medSignalVal} - The median broadPeak "signalVal" (see \href{https://genome.ucsc.edu/FAQ/FAQformat.html#format13}{UCSC specification}) of sample peaks in the RL region.
\item \code{medPVal} - The median broadPeak -log10 pval (see \href{https://genome.ucsc.edu/FAQ/FAQformat.html#format13}{UCSC specification}) of sample peaks in the RL region.
\item \code{medQVal} - The median broadPeak -log10 qval (adjusted pval) (see \href{https://genome.ucsc.edu/FAQ/FAQformat.html#format13}{UCSC specification}) of sample peaks in the RL region.
\item \code{avgSignalVal/avgPVal/avgQVal} - same as above, but with mean instead of median.
\item \code{medNumPeaks} - The median number of peaks for samples which overlapped with this RL Region.
\item \code{avgNumPeaks} - Same as above with mean instead of median.
\item \code{nSamples} - The number of samples which overlap with this peak.
\item \code{nModes} - The number of different R-loop mapping methods in which this RL region was found.
\item \code{nTissues} - The number of different tissues in which this RL region was found.
\item \code{ip_types} - The "IP types" (S9.6, dRNH, or None) in which this RL Region was found.
\item \code{nIPType} - The number of different "IP types" in which this RL region was found.
\item \code{pct_case} - The percentage of samples which overlapped with this RL region that were predicted "POS" by \code{\link[RLSeq:predictCondition]{RLSeq::predictCondition()}}.
\item \code{geneIDs} - The IDs of all genes overlapping with this RL Region.
\item \code{allGenes} - The Gene Symbols of all genes overlapping with this RL Region.
\item \code{mainGenes} - The Gene Symbols of all genes (which appear in pathway databases) overlapping with this RL Region.
\item \code{is_repeat} - A logical indicating whether the RL Region occurs in a repetitive genomic region as determined by repeat masker (and also peri-centromeric regions were included.)
}
}

\subsection{RL Regions anno - \link{rlregions_annot}}{

\code{rlregions_annot()} is a \code{tbl} with the following structure:\tabular{ll}{
   rlregion \tab annotation \cr
   All_RL4615 \tab CNA__3__4 \cr
   All_RL4615 \tab CNA__1__8 \cr
   All_RL4615 \tab G4Qexp__G4Q_PDS_NaK_GSE63874__95848 \cr
   ... \tab ... \cr
}


Column description:
\itemize{
\item \code{rlregion} - The RL Region ID corresponding to the \link{rlregions_meta} table.
\item \code{annotation} - The annotation which overlaps with the query RL region. Annotation
names follow the convention "DataBase__Type__ID" (see also \link{annotations}).
}
}

\subsection{RL Regions Counts - \link{rlregions_counts}}{

\code{rlregions_counts()} is a \code{SummarizedExperiment} with the following structure:
\itemize{
\item \code{dim}: 64418 532
\item \code{assays}
\itemize{
\item \code{cts} - Raw read counts of each RLBase sample within each RL Region.
\item \code{tpm} - The TPM normalization of \code{cts}
\item \code{vst} - The VST transform of \code{cts} (see also \code{\link[DESeq2:vst]{DESeq2::vst()}}).
}
\item \code{colData} - A \code{DataFrame} with structure:
}\tabular{lllllllrll}{
   experiment \tab label \tab strand_specific \tab paired_end \tab mode \tab prediction \tab discarded \tab numPeaks \tab bam \tab bam_avail \cr
   SRX113814 \tab POS \tab FALSE \tab FALSE \tab DRIVE \tab NEG \tab TRUE \tab 337 \tab ../RLBase-data/rlbase-data/rlpipes-out/bam/SRX113814/SRX113814_hg38.bam \tab TRUE \cr
   SRX113812 \tab POS \tab FALSE \tab FALSE \tab DRIP \tab NEG \tab TRUE \tab 78 \tab ../RLBase-data/rlbase-data/rlpipes-out/bam/SRX113812/SRX113812_hg38.bam \tab TRUE \cr
   SRX2675003 \tab POS \tab TRUE \tab FALSE \tab R-ChIP \tab POS \tab FALSE \tab 10331 \tab ../RLBase-data/rlbase-data/rlpipes-out/bam/SRX2675003/SRX2675003_hg38.bam \tab TRUE \cr
   ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \cr
}


Column description:
\itemize{
\item \code{experiment} - The RLBase sample ID
\item \code{label} - The label provided by the data submitters ("POS" or "NEG")
\item \code{strand_specific} - A logical indicating whether the data is stranded or not.
\item \code{paired_end} - A logical indicating whether the data is paired-end or not.
\item \code{prediction} - The predicted label from \code{\link[RLSeq:predictCondition]{RLSeq::predictCondition()}}. Can be "POS" or "NEG".
\item \code{discarded} - A logical indicating whether the sample was discarded during model building due to a mismatch with its label.
\item \code{numPeaks} - The number of peaks called from this sample.
\item \code{bam} - The relative path of the bam file for this sample.
\item \code{bam_avail} - A logical indicating whether the bam file was available.
}
}
}

\examples{

rlregions <- rlregions_meta()

rlregionsAnno <- rlregions_annot()

rlregionsCounts <- rlregions_counts()

}
