% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLHub-package.R
\docType{package}
\name{RLHub-package}
\alias{RLHub-package}
\alias{_PACKAGE}
\alias{RLHub-pacage}
\alias{annots_full_hg38}
\alias{annots_full_mm10}
\alias{annots_primary_hg38}
\alias{annots_primary_mm10}
\alias{feat_enrich_rlregions}
\alias{feat_enrich_samples}
\alias{fft_model}
\alias{gene_exp}
\alias{gs_signal}
\alias{prep_features}
\alias{rlbase_samples}
\alias{rlbps}
\alias{rlfs_res}
\alias{rlregions_annot}
\alias{rlregions_counts}
\alias{rlregions_meta}
\title{Accessing RLHub data}
\description{
RLHub is a convenient access point to the processed datasets available
within RLBase. In this document, we list the availabl accessor functions
which streamline access to the data.
}
\details{
The RLHub package provides a useful interface to the data stored within the
"RLHub" ExperimentHub database. The full manifest of available data is found
here:

\code{system.file("extdata", "metadata.csv", package = "RLHub")}
}
\examples{

## Access data-sets ##

# Manifest of samples in RLBase
dat <- rlbase_samples()

# Full hg38 annotations
dat <- annots_full_hg38()

# Full mm10 annotations
dat <- annots_full_mm10()

# Primary hg38 annotations
dat <- annots_primary_hg38()

# Primary mm10 annotations
dat <- annots_primary_mm10()

# Feature enrichment of RLRegions
dat <- feat_enrich_rlregions()

# Feature enrichment of RLBase samples
dat <- feat_enrich_samples()

# FFT-based sample classifier model
dat <- fft_model()

# Feature transformer for fft_model()
dat <- prep_features()

# Gene expression results from RLBase
dat <- gene_exp()

# "Gold-standard" R-loop signal for samples in RLBase
dat <- gs_signal()

# R-loop binding proteins data
dat <- rlbps()

# Results of RLFS analysis from RLBase
dat <- rlfs_res()

# R-loop regions, annotated for hg38 features
dat <- rlregions_annot()

# R-loop regions, with read counts (as a summarized experiment)
dat <- rlregions_counts()

# R-loop regions, metadata
dat <- rlregions_meta()

}
\author{
\strong{Maintainer}: Henry Miller \email{millerh1@uthscsa.edu} (\href{https://orcid.org/0000-0003-3756-3918}{ORCID})

}
\keyword{internal}
